made all the changes mentioned
