+++ core/package.json
@@ -25,7 +25,7 @@
     "@aws-sdk/credential-providers": "3.556.0",
     "@emotion/react": "11.11.3",
     "@emotion/styled": "11.11.0",
-    "@likeminds.community/feed-js": "1.19.0",
+    "@likeminds.community/feed-js": "1.20.0",
     "@mui/material": "5.15.10",
     "@mui/x-date-pickers": "7.22.2",
     "@testing-library/jest-dom": "5.14.1",
