+++ core/src/assets/scss/components/_lm-feed-upload-banner.scss
@@ -4,43 +4,42 @@
   left: 0;
   right: 0;
   z-index: 1000;
-  background: linear-gradient(to right, #4a90e2, #357abd);
-  padding: 16px;
+  background: var(--lm-feed-info-color);
+  padding: pxToRem(16);
   display: flex;
   align-items: center;
   justify-content: center;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
   transition: all 0.3s ease;
 
   &__container {
-    max-width: 1200px;
+    max-width: pxToRem(1200);
     width: 100%;
     margin: 0 auto;
     display: flex;
     align-items: center;
     justify-content: center;
-    gap: 16px;
+    gap: pxToRem(16);
   }
 
   &__text {
-    color: white;
-    font-size: 16px;
+    color: var(--lm-feed-light-color);
+    font-size: pxToRem(16);
     font-weight: 500;
-    margin-right: 16px;
+    margin-right: pxToRem(16);
   }
 
   &__retry-btn,
   &__cancel-btn {
-    padding: 8px 16px;
+    padding: pxToRem(8) pxToRem(16);
     border: none;
-    border-radius: 6px;
-    font-size: 14px;
+    border-radius: pxToRem(6);
+    font-size: pxToRem(14);
     font-weight: 500;
     cursor: pointer;
     transition: all 0.2s ease;
 
     &:hover {
-      transform: translateY(-1px);
+      transform: translateY(pxToRem(-1));
     }
 
     &:active {
@@ -49,43 +48,43 @@
   }
 
   &__retry-btn {
-    background-color: white;
-    color: #4a90e2;
-    margin-right: 12px;
+    background-color: var(--lm-feed-light-color);
+    color: var(--lm-feed-primary-color);
+    margin-right: pxToRem(12);
 
     &:hover {
-      background-color: #f5f5f5;
+      background-color: var(--lm-feed-light-color);
     }
   }
 
   &__cancel-btn {
-    background-color: rgba(255, 255, 255, 0.15);
-    color: white;
-    backdrop-filter: blur(8px);
+    background-color: var(--lm-feed-danger-color);
+    color: var(--lm-feed-light-color);
+    backdrop-filter: blur(pxToRem(8));
 
     &:hover {
-      background-color: rgba(255, 255, 255, 0.25);
+      background-color: var(--lm-feed-dark-color);
     }
   }
 
   // Error state
   &--error {
-    background: linear-gradient(to right, #e74c3c, #c0392b);
+    background: var(--lm-feed-danger-color);
 
     .lm-feed-upload-banner__retry-btn {
-      color: #e74c3c;
+      color: var(--lm-feed-light-color);
     }
   }
 
   // Loading animation for the uploading state
   &__loading {
     display: inline-block;
-    width: 20px;
-    height: 20px;
+    width: pxToRem(20);
+    height: pxToRem(20);
     margin-right: 12px;
-    border: 2px solid rgba(255, 255, 255, 0.3);
+    border: pxToRem(2) solid var(--lm-feed-dark-color);
     border-radius: 50%;
-    border-top-color: white;
+    border-top-color: var(--lm-feed-light-color);
     animation: spin 1s linear infinite;
   }
 }
