+++ core/src/components/lmDialogs/LMFeedDeleteDialogBox.tsx
@@ -61,4 +61,5 @@ const LMFeedDeleteDialogBox = ({ mode, onClose }: LMDeletePostDialogProps) => {
   );
 };
 
+
 export default LMFeedDeleteDialogBox;
