+++ core/src/components/lmTopicFeed/LMFeedTopicSelectedBlock.tsx
@@ -36,4 +36,5 @@ export const LMFeedTopicSelectedBlock = ({
       );
     }
   }
+  
 };
