+++ core/src/components/lmTopicFeed/LMFeedTopicSelectionTile.tsx
@@ -29,4 +29,5 @@ const LMFeedTopicSelectionTile: React.FC<LMFeedTopicSelectionTileProps> = memo(
   },
 );
 
+
 export default LMFeedTopicSelectionTile;
