+++ core/src/components/lmTopicFeed/LMFeedTopicsTile.tsx
@@ -13,4 +13,5 @@ const LMFeedTopicsTile = ({ topic }: LMFeedTopicTileInterface) => {
   );
 };
 
+
 export default LMFeedTopicsTile;
